update docs and put controllers in order
