+++ views/documentation.ejs
@@ -121,7 +121,7 @@
 
                  <h4>Update Password</h4>
             <ul>
-                <li>Update password<strong> - /update/password [POST]</strong>
+                <li>Update password<strong> - /store-admin/update/password [POST]</strong>
                     <h6>Body: </h6>
                     <ul class="alert alert-primary">
                         <li>old_password</li>
